#19 [refactor] : 마이 페이지 폴더 구조 변경
